requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
psycopg==3.1.18
elasticsearch==8.12.0
psycopg2-binary==2.9.11
requests==2.32.5
python-dotenv==1.0.1
PyJWT==2.9.0
whisperlivekit==0.2.16
faster-whisper==1.2.1
mediapipe==0.10.20
opencv-python==4.10.0.84
numpy==1.26.4
django-redis==6.0.0
celery==5.6.0
redis==6.4.0 
ruff==0.14.8

# langgraph
langchain==1.1.0
langchain-community==0.4.1
langchain-openai==1.1.0
langgraph==1.0.4
langgraph-checkpoint-redis==0.2.1

# media 분석용
pandas==2.3.3
scikit-learn==1.6.1
joblib==1.5.2
